# api/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
        "/api/branches": {
            "get": {
                "summary": "Returns all branches",
                "tags": [
                    "Branches"
                ],
                "responses": {
                    "200": {
                        "description": "List of all branches",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Branch"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new branch",
                "tags": [
                    "Branches"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Branch created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/branches/{id}": {
            "get": {
                "summary": "Get a branch by ID",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "put": {
                "summary": "Update a branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Branch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Branch updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Branch"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a branch",
                "tags": [
                    "Branches"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Branch deleted successfully"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
...
        "/api/headquarters": {
            "get": {
                "summary": "Returns all headquarters",
                "tags": [
                    "Headquarters"
                ],
                "responses": {
                    "200": {
                        "description": "List of all headquarters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Headquarters"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new headquarters",
                "tags": [
                    "Headquarters"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Headquarters created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/headquarters/{id}": {
            "get": {
                "summary": "Get a headquarters by ID",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Headquarters found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "put": {
                "summary": "Update a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Headquarters updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Headquarters deleted successfully"
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Branch": {
                "type": "object",
                "required": [
                    "branchId",
                    "name",
                    "headquartersId"
                ],
                "properties": {
                    "branchId": {
                        "type": "integer",
                        "description": "The unique identifier for the branch"
                    },
                    "headquartersId": {
                        "type": "integer",
                        "description": "The ID of the headquarters this branch belongs to"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the branch"
                    },
                    "address": {
                        "type": "string",
                        "description": "Physical address of the branch"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the branch"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the branch"
                    }
                }
            },
...
            "Headquarters": {
                "type": "object",
                "required": [
                    "headquartersId",
                    "name"
                ],
                "properties": {
                    "headquartersId": {
                        "type": "integer",
                        "description": "The unique identifier for the headquarters"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the headquarters"
                    },
                    "address": {
                        "type": "string",
                        "description": "Main office address of the headquarters"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the headquarters"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the headquarters"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the headquarters"
                    }
                }
            },
...
        }
    },
...
}